refac(campagne) - Extract favoris in new service
