+++ src/jdRoll/controller/public_campagn.php
@@ -64,13 +64,13 @@ $publicCampagneController->post('/favoris', function(Request $request) use($app)
             $statut = $request->get("statut");
 
             if($statut == 0) {
-                $app["campagneService"]->removeFavoris($campagne, $joueur);
+                $app["favorisService"]->removeFavoris($campagne, $joueur);
             } else {
-                $app["campagneService"]->addFavoris($campagne, $joueur);
+                $app["favorisService"]->addFavoris($campagne, $joueur);
             }
             return "";
         })->bind("favoris");
-		
+
 
 $app->mount('/campagne', $publicCampagneController);
